Add a test for the status component
